// SDKs/amazonaws.com/devicefarm/2015-06-23/typescript/src/sdk/models/shared/customerartifactpaths.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";


// CustomerArtifactPaths
/** 
 * <p>A JSON object that specifies the paths where the artifacts generated by the customer's tests, on the device or in the test environment, are pulled from.</p> <p>Specify <code>deviceHostPaths</code> and optionally specify either <code>iosPaths</code> or <code>androidPaths</code>.</p> <p>For web app tests, you can specify both <code>iosPaths</code> and <code>androidPaths</code>.</p>
**/
export class CustomerArtifactPaths extends SpeakeasyBase {
  @Metadata({ data: "json, name=androidPaths" })
  androidPaths?: string[];

  @Metadata({ data: "json, name=deviceHostPaths" })
  deviceHostPaths?: string[];

  @Metadata({ data: "json, name=iosPaths" })
  iosPaths?: string[];
}